chore: organize files and folders
