+++ miso-hungry/src/components/Result.jsx
